for dhclient use g_spawn_sync instead of g_spawn_async
For some (unknown) reason, g_spawn_async() is not starting dhclient
so that a dhcp NIC can be configured.  However, siply using
g_spawn_sync() works.  This was the only use of g_spawn_async().

Note: There is no problem using sync instead of async since dhclient will
disconnect and put itself in the background once the network is started.